requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-amazon-sp-api==0.9.4
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==1.10.17
itsdangerous==2.2.0
python-multipart==0.0.9
